Refactoring Integracion
