+++ jobs/integracionSipsJob.ts
@@ -1,7 +1,7 @@
 import * as agendaCtrl from './../modules/turnos/controller/agendasCacheController';
 
 function run() {
-    agendaCtrl.getAgendaSips();
+    agendaCtrl.integracionSips();
 }
 
 export = run;
